+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Initial release.
 
 ### New Commands
 
-**find**: The find command returns exit code 0 when the given input contains at least one image reference that satisfy the given conditions, non-null otherwise
+**contains**: The contains command returns exit code 0 when the given input contains at least one image reference that satisfy the given conditions, non-null otherwise
 
 ### New Formats